refactor testing and some minor improvment
